Suggest using URNs for eventID and introduced eventGenerator
In large systems, non namespaced eventIds will require that events are wrapped in an envelope to avoid collisions. No guarantee that a producer doesn't use a counter for its event ID. The urn provides a standards based mechanism of making the ID unique while still allowing the producer the flexibility to use the a commit ID.

Similarly without some context provided by the eventGenerator the consumer will require some out of band knowledge of the meaning of the event. Having a URI (urn or stable url) as an eventGenerator provides a mechanism by which the consumer can discover the meaning of data elements, without having to infer context from the eventType. Abstract eventTypes like "document.updated", "asset.created" may have different significance in different contexts.